requirements.txt
streamlit
wikipedia
sentence-transformers
faiss-cpu
transformers
torch
numpy
